refactor: apis/user-point-ranking 폴더구조 변경
