+++ src/components/layouts/with-side/profile.tsx
@@ -3,8 +3,8 @@
 import ComponentFrame from '@/components/common/component-frame';
 import { useCurrentUserInfoStore } from '@/lib/store/useCurrentUserInfoStore';
 import { useIsLoginModalOpenStore } from '@/lib/store/useIsLoginModalOpenStore';
-import { getUserPointRanking } from '@/services/apis/user-point-event';
-import { UserPointRankingDto } from '@/services/apis/user-point-event/dto';
+import { getUserPointRanking } from '@/services/apis/user-point-event/user-point-event.api';
+import { UserPointRankingDto } from '@/services/apis/user-point-event/user-point-event.type';
 import { getUserInfo } from '@/services/apis/user/user.api';
 import Image from 'next/image';
 import Link from 'next/link';
